edge-sensor-network/services/edge-node/requirements.txt
Flask==2.0.1
paho-mqtt==1.5.1
pyyaml==5.4.1
requests==2.25.1
threading==3.9.7